+++ server/budgettracker.business/Serialization/Converters/BudgetSerializationConverter.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using budgettracker.common;
+using budgettracker.common.Models;
 
 namespace budgettracker.business.Serialization.Converters
 {
